Added support for Travis CI Started working on testing
