+++ .travis.yml
@@ -1,0 +1,8 @@
+sudo: required
+language: node_js
+node_js: node
+install: npm install
+addons:
+  apt:
+    packages:
+    - arp-scan
