+++ package.json
@@ -1,10 +1,12 @@
 {
   "name": "monitroom",
-  "version": "0.0.0",
+  "version": "0.0.1",
   "private": true,
   "scripts": {
-    "start": "node ./bin/www"
+    "start": "node ./bin/www",
+    "test": "./node_modules/mocha/bin/mocha"
   },
+  "license": "MIT",
   "dependencies": {
     "arpscan": "^0.1.0",
     "body-parser": "~1.15.1",
@@ -17,5 +19,9 @@
     "serve-favicon": "~2.3.0",
     "speedtest-net": "^1.2.7",
     "winston": "^2.2.0"
+  },
+  "devDependencies": {
+    "chai": "^3.5.0",
+    "mocha": "^3.0.2"
   }
 }
